Refactor cardano-cryto.js shelley functionality
