+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cardano-crypto.js",
-  "version": "5.3.6-rc.2",
+  "version": "5.3.6-rc.3",
   "description": "input-output-hk/cardano-crypto compiled to pure javascript using Emscripten",
   "main": "index.js",
   "scripts": {
